+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/meta/IsErrorNode.java
@@ -8,7 +8,8 @@ import org.enso.interpreter.runtime.type.TypesGen;
 @BuiltinMethod(
     type = "Meta",
     name = "is_error",
-    description = "Checks if the argument is an error.")
+    description = "Checks if the argument is an error.",
+    autoRegister = false)
 public class IsErrorNode extends Node {
   boolean execute(@AcceptsError Object value) {
     return TypesGen.isDataflowError(value);